Adding adbeat bot useragents
Thought this one seemed easy enough to attempt on my own.

Fixes #1204.

I added the unique useragents from the list in #1204 to the test suite
and added necessary JSON to make the tests pass.  I believe I
approached this the right way, please look it over and let me know if
anything should be done differently.