Adding UC Browser versions and new devices for #1223
This adds UC Browser versions 11.1 and 11.2 (which all of the
useragents in #1223 are). I also added a lot of new devices for these
strings (a couple of them had a device identifier that wasn’t very
specific, so I left those ones as “general Mobile Phone”).  Test added
for each agent in that ticket, all tests pass.