POC of device expansion
Discussed more in #1239.  Consolidating patterns that only differ by a
device identifier, using a “devices” array that contains the name of
the device as the key (this would have been the “device” key’s value
previously), and the pattern to use to match that device as the value.

The pattern will replace the token `#DEVICE#` that is in the match
string for each device listed in the `devices` array.

Currently, this only works in children rules, but I think it could be
expanded to be used in the parent as well, if that makes sense.

I’ve added some source tests and converted two UA JSON files (Android
browser 4.1 and 4.2) as an example of what this could look like (and to
test that it functions as expected and all UA tests still pass).