Condensing devices for UC Browser 7
While doing the condensing, I noticed that there was a pattern that
seemed like it should have a device linked, but didn’t:

`S8500 UCWEB6.0\/UC Browser#MAJORVER#.#MINORVER#*`

I added the device, based on data from here:

http://www.gsmarena.com/samsung_s8500_wave-3146.php

And noticed that no tests failed after adding the device. When the
pattern was added, it doesn’t seem like there was a test to go with it.

Dug around in our logs a bit and found this UA that has the same prefix
as the pattern:

>S8500 UCWEB6.0/UC Browser7.7.0.81

So added it as a test (to the issue-000-mobiles test file), which
passes.