condensing facebook app devices
Another file with high coverage.

There are a couple of patterns in this file that I wanted to address.

One is an exact duplicate, so I removed it:
https://codecov.io/gh/jaydiablo/browscap/src/1ca764c275e7c01b2331e517a56
bd839abbc5428/resources/user-agents/apps/facebook/facebook-app.json#L348
 (duplicate of this one:
https://codecov.io/gh/jaydiablo/browscap/src/1ca764c275e7c01b2331e517a56
bd839abbc5428/resources/user-agents/apps/facebook/facebook-app.json#L362
)

Then there was this one:
https://codecov.io/gh/jaydiablo/browscap/src/1ca764c275e7c01b2331e517a56
bd839abbc5428/resources/user-agents/apps/facebook/facebook-app.json#L990

It’s almost a duplicate of the one above it:
https://codecov.io/gh/jaydiablo/browscap/src/1ca764c275e7c01b2331e517a56
bd839abbc5428/resources/user-agents/apps/facebook/facebook-app.json#L982

But differs slightly in the ending of the pattern
(`Version/*Safari/*FB_IAB/*` vs. `Version/*Safari/*FBAV/*`).

The test useragent that hits the covered pattern here is this one:

>Mozilla/5.0 (Linux; U; Android 4.2.2; uk-ua; M6 Build/JDQ39)
AppleWebKit/534.30 (KHTML, like Gecko) Version/4.0 Safari/534.30
[FB_IAB/FB4A;FBAV/31.0.0.20.13;]

Located in this fixture:
https://github.com/browscap/browscap/blob/master/tests/fixtures/issues/i
ssue-635.php (too large to link to the actual line, but it’s on line
22848).

If the `IAB` agent is removed, that test fails, this seems to be due to
a length/position in the file issue, as the pattern that it ends up
selecting is the same length as the `FBAV` pattern that should be
selected.

Here’s the pattern selected when I remove the `IAB` pattern:

`mozilla/5.0 (*linux*android?4.2* build/*) applewebkit*
(*khtml*like*gecko*) version/*safari/*fb_iab/*` (101, 88, 283126)

Here’s what the `IAB` pattern looks like fully assembled (the one that
browscap uses currently):

`mozilla/5.0 (*linux*android?4.2*m6 build/*) applewebkit*
(*khtml*like*gecko*) version/*safari/*fb_iab/*` (103, 90, 282656)

And here’s the pattern that it **should** be using if the `IAB` pattern
is removed:

`mozilla/5.0 (*linux*android?4.2*m6 build/*) applewebkit*
(*khtml*like*gecko*) version/*safari/*fbav/*` (101, 88, 282466)

The numbers on the end are: (strlen of pattern, strlen of pattern with
asterisk/question mark removed, position in browscap.ini)

This is probably related to the issues described in
https://github.com/browscap/browscap-php/issues/179

The other devices for that particular pattern don’t have this issue
because their “match” pushes their length longer, thus increasing their
priority.

Just a heads up, not planning to touch that pattern at all in this PR.
Perhaps that generic catch-all pattern should be modified to use `FBAV`
instead of `FB_IAB` to be more consistent with the other patterns in
the file? (thus shortening its length to be shorter than even a 1
character device code)