Addressing review comments, refactored Processor to split out into multiple methods
This adds scalar type hints to all arguments and return values where
possible (in the pattern coverage changes, not the entire code base)
which should address most of the comments from @asgrim.

Added `@return void` where necessary (no `: void` yet since it’s not
minimum 7.1). Had to remove the `phpdoc_no_empty_return` rule from the
PHPCSFixer config in order for these to stick once the fixer is run.

Refactored the coverage processor to use multiple methods instead of a
large switch statement.  I think this cleans up the logic quite a bit
and makes it easier to understand.

Also removing the composer.lock gitignore entry, which adds the current
lock file to the repo.