Some more missing parameter types and return types
I also decided to add return types where applicable to
Browscap\Data\Expander.php and trun on the strict_types for that file.

This did require casting the value sent to the `trimProperty` method to
a string, as it would fail the strict type check otherwise.  If I
remove the parameter type hint for that method, it actually causes a
boolean to be sent to `trim()` (I assume this is because PHP doesn’t do
any implicit casting when the strict_types declaration is done).
Casting before sending to that method fixes in either case.