Adding Firefox on iOS versions/platforms/devices
This fixes #1272.

In addition to the UA in the referenced ticket I also dug through our
server logs and collected a variety of Firefox on iOS useragents that
spanned different versions/iOS versions/devices.

This resulted in me adding a number of Firefox iOS versions, a few
platforms to the Firefox 1.x file, and iPod Touch as a device for both
Firefox iOS files.