Adding 1password versions, platforms
This fixes #1306.

I wanted to add some other test cases, but not finding much in logs (an
Android example would be nice, but it’s possible that it uses a generic
UA and passes along a different header, based on what I saw here:
https://whichbrowser.net/data/browserids/index.html).

There is another UA in #1306 I could add, but it requires that #1312 is
merged first since it’s on iOS 10.3 (which will need to be added to the
1password.json file as well).  Depending on the order of merging, I can
update this or #1312 to add 10.3 support.

I intentionally skipped version 4.4 because it isn’t listed on agile
bits’ changelog: https://app-updates.agilebits.com/product_history/OPI4