Condensing devices in/cleaning up Android Browser files
This should wrap up device expansion for the Android Browser files. The
4.0 file was certainly the one with the most devices that were able to
be combined.

I did a little bit of cleanup to that file in particular, adding some
devices that were set as “general Mobile Device” or that didn’t have a
device at all.  There were a few patterns that were eligible for being
condensed into another pattern, but the device identifier was too
ambiguous to be able to determine what it was.  Ultimately I removed
these patterns entirely (I recall a few device ids: `A*`, `OYO`). I
didn’t see any particular reason for these patterns to be in the file
with no device defined, as they should just fall back to a default
pattern that would have “general Mobile Phone” as the device anyhow.
Unfortunately, none of the patterns like this (including the ones I
added devices for) are covered by tests.

The one test case I changed was because I moved a couple of patterns
that looked like they were meant for different versions of Android
Browser into their appropriate files, and replaced the hard-coded “5.0”
or “5.1” after the “Version/“ part with the respective version tokens.
One test covered these patterns and claimed that the android browser
version should be “4.0” when the “Version/5.0” part of the UA seemed to
claim otherwise (let me know if this is incorrect).