Adding some devices/versions of Firefox on FirefoxOS
Fixes #1350

Found some patterns that weren’t matching to FirefoxOS like they should
have been.

I added Firefox OS 2.5, which I’ve marked as Beta. Looks like it was
the next planned release of Firefox OS, but the project was cancelled
before it was released. I assume that the versions of Firefox on the
OPENC phone I’m seeing are due to this beta release, but they could be
fake.

https://en.wikipedia.org/wiki/Firefox_OS#Release_history