Adding test coverage for Android Browser 4.0
This references #885 (loosely).

I was curious what the overlapping patterns looked like (using the
python script posted in that ticket) currently, so ran it against a
full INI build.

I noticed that there was a lot of device related patterns in Android
4.0 (other files too, but I started here) that came up.  Before messing
with the patterns too much, I wanted to see if it was possible to boost
coverage by pulling some useragents from our logs.

This is basically an attempt at that, plus some clean up of the devices
encountered when adding the tests (most of it was adding a Device Name
where there wasn’t one, or changing the “Pointing Method” to
“touchscreen” instead of “unknown”).

I’ll add some inline comments on the less obvious changes.

I did end up reducing some of the overlapping patterns in the Android
4.0 file (and fixing a mis-detection here and there).