Adding Samsung Galaxy A3, A5, A7 (2016) devices to various browsers/apps
Fixes #1392

I added the A3/A5/A7 devices to different browsers that we saw show up
in our logs, with tests for different versions/platforms based on what
we were seeing.

I only added Android 5.1 and 6.0 for these devices, as that’s mostly
all I saw in our logs, and that seems to be the versions of Android
that these devices support.

Also bumped the android versions on the fallback pattern in the version
files I touched to include 5.1 and 6.0.