Adding some missing macOS versions to Opera
Fixes #2379. This fixes situations where Opera was being detected as Chrome on some versions of macOS. Added additional useragents found in our logs to improve test coverage.

There was one test in the suite (1953-BGU) that was actually an Opera UA being detected as Chrome, I removed it from the suite.