Adding missing versions, adding DuckDuckGo browser
Fixes #2396.  This is mostly a version update for various browsers that had no version from our Nov logs. I added Android and iOS versions where necessary as well.

Also noticed a number of DuckDuckGo UAs that appear to go with a DuckDuckGo browser available for Mac, iOS and Android, so added a version-less browser and patterns for those to be detected now.

Apparently Chrome 83 dropped support for Android 4.4, so I split the Chrome and Android Webview files at 82/83 and removed Android 4.4 from the 83+ files.